[Template] Unify Header component
